myapplication/requirements.txt
flask
cherrypy
python-whois